rename package to `tgen`
