+++ consensus-types/blocks/BUILD.bazel
@@ -40,7 +40,7 @@ go_test(
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "//testing/util:go_default_library",
-        "//testing/util/tgenerator:go_default_library",
+        "//testing/util/tgen:go_default_library",
         "@com_github_ferranbt_fastssz//:go_default_library",
     ],
 )
