+++ consensus-types/interfaces/BUILD.bazel
@@ -30,6 +30,6 @@ go_test(
         "//proto/prysm/v1alpha1:go_default_library",
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
-        "//testing/util/tgenerator:go_default_library",
+        "//testing/util/tgen:go_default_library",
     ],
 )
