+++ testing/util/tgen/generator.go
@@ -1,4 +1,4 @@
-package tgenerator
+package tgen
 
 import (
 	"math/rand"
